+++ software_esp32/include/compile_time.h
@@ -80,4 +80,3 @@
 #endif /* COMPILE_TIME_H_ */
 
 
-