+++ software_stm32/include/helper.h
@@ -37,48 +37,11 @@
 
 *END************************************************************************/
 
+#include <Arduino.h>
 
 #pragma once
 
-#include <Arduino.h>
-#include "VdmNet.h"
-#include "VdmConfig.h"
-#include "globals.h"
-#include "stmApp.h"
-
-class CWeb
-{
-public:
-  CWeb();
-  void handleFileUpload();
-  void handleFlash();
-  void handleFileDelete();
-  void handleListFiles();
-  String makePage(String title, String contents);
-  void postValvesPos();
-  String getValvesStatus();
-  String getTempsStatus(VDM_TEMPS_CONFIG tempsConfig);
-  String getVoltsStatus(VDM_VOLTS_CONFIG voltsConfig); 
-  String getNetInfo(VDM_NETWORK_INFO networkInfo); 
-  String getNetConfig (VDM_NETWORK_CONFIG netConfig);
-  String getSSIDInfo();
-  String getProtConfig (VDM_PROTOCOL_CONFIG protConfig);
-  String getValvesConfig (VDM_VALVES_CONFIG valvesConfig);
-  String getMotorConfig (MOTOR_CHARS motorConfig);
-  String getValvesControlConfig (VDM_VALVES_CONTROL_CONFIG valvesControlConfig);
-  String getTempsConfig (VDM_TEMPS_CONFIG tempsConfig);
-  String getVoltsConfig (VDM_VOLTS_CONFIG voltsConfig);
-  String getTempSensorsID();
-  String getVoltSensorsID();
-  String getSysDynInfo();
-  String getSysInfo();
-  String getSysConfig (VDM_SYSTEM_CONFIG sysConfig);
-  String getMsgConfig (VDM_MSG_CONFIG msgConfig);
-  String getFSDir();
-  String getStmUpdStatus();
-  bool findIdInValve (uint8_t idx);
-  bool getControlActive();
-};
-
-extern CWeb Web;
-
+void replace (char* buffer,uint16_t size,char find, char with);
+char* findCharInString (char c,char* pString,size_t size);
+char* copyStringUntilChar (char* pString,char* buffer,char c,size_t pSize,size_t bSize);
+size_t strlcat(char *dst, const char *src, size_t dsize);